remove read&write operation into another file
